Update query for alert stock in home

+++ app/src/main/java/com/lamz/trackinv/presentation/ui/screen/add/AddProductScreen.kt
@@ -232,7 +232,7 @@ fun AddProductContent(
                     val barang = BarangModel(
                         idBarang,
                         namaBarang,
-                        stokBarang,
+                        stokBarang.toInt(),
                         hargaBeli,
                         hargaJual
                     )
